UPDATE: Move ewallet-modal and converted to Context Modal
